+++ e2e/tests-dfx/generate.bash
@@ -4,8 +4,6 @@ load ../utils/_
 
 setup() {
     standard_setup
-
-    dfx_new hello
 }
 
 teardown() {
@@ -15,6 +13,7 @@ teardown() {
 }
 
 @test "dfx generate creates files" {
+    dfx_new hello
     dfx_start
     dfx canister create --all
     dfx build
@@ -30,6 +29,7 @@ teardown() {
 }
 
 @test "dfx generate creates only JS files" {
+    dfx_new hello
     jq '.canisters.hello_backend.declarations.bindings=["js"]' dfx.json | sponge dfx.json
 
     dfx_start
@@ -47,6 +47,7 @@ teardown() {
 }
 
 @test "dfx generate creates only TS files" {
+    dfx_new hello
     jq '.canisters.hello_backend.declarations.bindings=["ts"]' dfx.json | sponge dfx.json
 
     dfx_start
@@ -64,6 +65,7 @@ teardown() {
 }
 
 @test "dfx generate creates only JS & TS files" {
+    dfx_new hello
     jq '.canisters.hello_backend.declarations.bindings=["js", "ts"]' dfx.json | sponge dfx.json
 
     dfx_start
@@ -81,6 +83,7 @@ teardown() {
 }
 
 @test "dfx generate creates only DID files" {
+    dfx_new hello
     jq '.canisters.hello_backend.declarations.bindings=["did"]' dfx.json | sponge dfx.json
 
     dfx_start
@@ -98,6 +101,7 @@ teardown() {
 }
 
 @test "dfx generate does not create any files" {
+    dfx_new hello
     jq '.canisters.hello_backend.declarations.bindings=[]' dfx.json | sponge dfx.json
 
     dfx_start
@@ -115,6 +119,7 @@ teardown() {
 }
 
 @test "dfx generate succeeds with an encrypted identity without input" {
+    dfx_new hello
     dfx_start
     dfx canister create --all
 
@@ -122,4 +127,41 @@ teardown() {
     assert_command dfx identity use alice
     
     assert_command timeout 30s dfx generate
+}
+
+@test "dfx generate does not require canister IDs for non-Motoko canisters" {
+    dfx_new_rust hello
+    assert_command dfx generate
+}
+
+@test "dfx generate requires canister IDs for Motoko canisters" {
+    dfx_new hello
+    assert_command_fail dfx generate
+    assert_contains "Please create canister 'hello_backend' before generating."
+}
+
+@test "dfx generate requires canister IDs for dependees of Motoko canister" {
+    dfx_new hello
+    dfx_start
+    jq '.canisters.hello_backend.dependencies[0]="dependee"' dfx.json | sponge dfx.json
+    jq '.canisters.dependee.type="assets"' dfx.json | sponge dfx.json
+    jq '.canisters.dependee.source=[]' dfx.json | sponge dfx.json
+    cat dfx.json
+    dfx canister create --all
+
+    # generate fails if Motoko canister itself is not created
+    dfx canister stop hello_backend
+    dfx canister delete hello_backend --no-withdrawal
+    assert_command_fail dfx generate
+    assert_contains "Please create canister 'hello_backend' before generating."
+    assert_command dfx canister create hello_backend
+
+    # generate fails if a dependee is not created
+    dfx canister stop dependee
+    dfx canister delete dependee --no-withdrawal
+    assert_command_fail dfx generate
+    assert_contains "Please create canister 'dependee' before generating."
+    assert_command dfx canister create dependee
+
+    assert_command dfx generate
 }