+++ src/dfx-core/src/error/foundation.rs
@@ -4,4 +4,7 @@ use thiserror::Error;
 pub enum FoundationError {
     #[error("Cannot find home directory (no HOME environment variable).")]
     NoHomeInEnvironment(),
+
+    #[error("Failed to identify currently running executable: {0}")]
+    NoCurrentExe(std::io::Error),
 }