+++ src/dfx-core/src/lib.rs
@@ -4,3 +4,6 @@ pub mod foundation;
 pub mod fs;
 pub mod identity;
 pub mod json;
+pub mod network;
+pub mod process;
+pub mod util;