+++ src/dfx/src/config/cache.rs
@@ -1,12 +1,10 @@
 use crate::config::dfx_version;
-use crate::lib::error::{CacheError, DfxError, DfxResult};
+use crate::lib::error::CacheError;
 use crate::util;
 #[cfg(windows)]
 use dfx_core::config::directories::project_dirs;
 
-use anyhow::{bail, Context};
 use dfx_core::foundation::get_user_home;
-use fn_error_context::context;
 use indicatif::{ProgressBar, ProgressDrawTarget};
 use rand::distributions::Alphanumeric;
 use rand::{thread_rng, Rng};
@@ -23,11 +21,11 @@ const EXEC_READ_USER_ONLY_PERMISSION: u32 = 0o500;
 pub trait Cache {
     fn version_str(&self) -> String;
     fn is_installed(&self) -> Result<bool, CacheError>;
-    fn install(&self) -> DfxResult;
-    fn force_install(&self) -> DfxResult;
-    fn delete(&self) -> DfxResult;
-    fn get_binary_command_path(&self, binary_name: &str) -> DfxResult<PathBuf>;
-    fn get_binary_command(&self, binary_name: &str) -> DfxResult<std::process::Command>;
+    fn install(&self) -> Result<(), CacheError>;
+    fn force_install(&self) -> Result<(), CacheError>;
+    fn delete(&self) -> Result<(), CacheError>;
+    fn get_binary_command_path(&self, binary_name: &str) -> Result<PathBuf, CacheError>;
+    fn get_binary_command(&self, binary_name: &str) -> Result<std::process::Command, CacheError>;
 }
 
 pub struct DiskBasedCache {
@@ -52,22 +50,21 @@ impl Cache for DiskBasedCache {
         is_version_installed(&self.version_str())
     }
 
-    fn install(&self) -> DfxResult {
+    fn install(&self) -> Result<(), CacheError> {
         install_version(&self.version_str(), false).map(|_| {})
     }
-    fn force_install(&self) -> DfxResult {
+    fn force_install(&self) -> Result<(), CacheError> {
         install_version(&self.version_str(), true).map(|_| {})
     }
-
-    fn delete(&self) -> DfxResult {
+    fn delete(&self) -> Result<(), CacheError> {
         delete_version(&self.version_str()).map(|_| {})
     }
 
-    fn get_binary_command_path(&self, binary_name: &str) -> DfxResult<PathBuf> {
+    fn get_binary_command_path(&self, binary_name: &str) -> Result<PathBuf, CacheError> {
         get_binary_path_from_version(&self.version_str(), binary_name)
     }
 
-    fn get_binary_command(&self, binary_name: &str) -> DfxResult<std::process::Command> {
+    fn get_binary_command(&self, binary_name: &str) -> Result<std::process::Command, CacheError> {
         binary_command_from_version(&self.version_str(), binary_name)
     }
 }
@@ -87,9 +84,7 @@ pub fn get_cache_root() -> Result<PathBuf, CacheError> {
         None => project_dirs()?.cache_dir().to_owned(),
     };
     if !p.exists() {
-        if let Err(_e) = std::fs::create_dir_all(&p) {
-            return Err(CacheError::CreateCacheDirectoryFailed(p));
-        }
+        dfx_core::fs::create_dir_all(&p).map_err(CacheError::CreateCacheDirectoryFailed)?;
     } else if !p.is_dir() {
         return Err(CacheError::FindCacheDirectoryFailed(p));
     }
@@ -102,9 +97,7 @@ pub fn get_bin_cache_root() -> Result<PathBuf, CacheError> {
     let p = get_cache_root()?.join("versions");
 
     if !p.exists() {
-        if let Err(_e) = std::fs::create_dir_all(&p) {
-            return Err(CacheError::CreateCacheDirectoryFailed(p));
-        }
+        dfx_core::fs::create_dir_all(&p).map_err(CacheError::CreateCacheDirectoryFailed)?;
     } else if !p.is_dir() {
         return Err(CacheError::FindCacheDirectoryFailed(p));
     }
@@ -121,36 +114,29 @@ pub fn is_version_installed(v: &str) -> Result<bool, CacheError> {
     get_bin_cache(v).map(|c| c.is_dir())
 }
 
-pub fn delete_version(v: &str) -> DfxResult<bool> {
+pub fn delete_version(v: &str) -> Result<bool, CacheError> {
     if !is_version_installed(v).unwrap_or(false) {
         return Ok(false);
     }
 
     let root = get_bin_cache(v)?;
-    std::fs::remove_dir_all(&root).with_context(|| {
-        format!(
-            "Failed to remove bin cache root {}.",
-            root.to_string_lossy()
-        )
-    })?;
+    dfx_core::fs::remove_dir_all(&root)?;
 
     Ok(true)
 }
 
-#[context("Failed to install binary cache for version '{}'.", v)]
-pub fn install_version(v: &str, force: bool) -> DfxResult<PathBuf> {
+pub fn install_version(v: &str, force: bool) -> Result<PathBuf, CacheError> {
     let p = get_bin_cache(v)?;
     if !force && is_version_installed(v).unwrap_or(false) {
         return Ok(p);
     }
 
-    if Version::parse(v).with_context(|| format!("Failed to parse version string {}.", v))?
+    if Version::parse(v).map_err(|e| CacheError::MalformedSemverString(v.to_string(), e))?
         == *dfx_version()
     {
         // Dismiss as fast as possible. We use the current_exe variable after an
         // expensive step, and if this fails we can't continue anyway.
-        let current_exe =
-            std::env::current_exe().context("Failed to identify currently running executable.")?;
+        let current_exe = dfx_core::foundation::get_current_exe()?;
 
         let b: Option<ProgressBar> = if atty::is(atty::Stream::Stderr) {
             let b = ProgressBar::new_spinner();
@@ -167,114 +153,69 @@ pub fn install_version(v: &str, force: bool) -> DfxResult<PathBuf> {
             .take(12)
             .map(|byte| byte as char)
             .collect();
-        let temp_p = get_bin_cache(&format!("_{}_{}", v, rand_string))
-            .context("Failed to get temporary bin cache path.")?;
-        std::fs::create_dir(&temp_p).with_context(|| {
-            format!(
-                "Failed to create temporary bin cache dir {}.",
-                temp_p.to_string_lossy()
-            )
-        })?;
+        let temp_p = get_bin_cache(&format!("_{}_{}", v, rand_string))?;
+        dfx_core::fs::create_dir_all(&temp_p)?;
 
         let mut binary_cache_assets =
-            util::assets::binary_cache().context("Failed to get asset binary cache.")?;
+            util::assets::binary_cache().map_err(CacheError::ReadBinaryCacheStoreFailed)?;
         // Write binaries and set them to be executable.
         for file in binary_cache_assets
             .entries()
-            .context("Failed to get binary cache archive entires.")?
+            .map_err(CacheError::ReadBinaryCacheEntriesFailed)?
         {
-            let mut file = file.context("Failed to get binary cache archive entry.")?;
+            let mut file = file.map_err(CacheError::ReadBinaryCacheEntryFailed)?;
 
             if file.header().entry_type().is_dir() {
                 continue;
             }
-            file.unpack_in(temp_p.as_path())
-                .context("Failed to unpack archive asset.")?;
+            dfx_core::fs::tar_unpack_in(temp_p.as_path(), &mut file)?;
             // On *nix we need to set the execute permission as the tgz doesn't include it
             #[cfg(unix)]
             {
-                let full_path = temp_p.join(file.path().context("Failed to get file path.")?);
-                let mut perms = std::fs::metadata(full_path.as_path())
-                    .with_context(|| {
-                        format!(
-                            "Failed to get file metadata for {}.",
-                            full_path.to_string_lossy()
-                        )
-                    })?
-                    .permissions();
+                let archive_path = dfx_core::fs::get_archive_path(&file)?;
+                let full_path = temp_p.join(archive_path);
+                let mut perms = dfx_core::fs::read_permissions(full_path.as_path())?;
                 perms.set_mode(EXEC_READ_USER_ONLY_PERMISSION);
-                std::fs::set_permissions(full_path.as_path(), perms).with_context(|| {
-                    format!(
-                        "Failed to set file permissions for {}.",
-                        full_path.to_string_lossy()
-                    )
-                })?;
+                dfx_core::fs::set_permissions(full_path.as_path(), perms)?;
             }
         }
 
         // Copy our own binary in the cache.
         let dfx = temp_p.join("dfx");
-        std::fs::write(
-            &dfx,
-            std::fs::read(current_exe).context("Failed to read currently running executable.")?,
-        )
-        .with_context(|| {
-            format!(
-                "Failed to copy running binary {} to cache.",
-                dfx.to_string_lossy()
-            )
-        })?;
+        dfx_core::fs::write(&dfx, dfx_core::fs::read(&current_exe)?)?;
         // On *nix we need to set the execute permission as the tgz doesn't include it
         #[cfg(unix)]
         {
-            let mut perms = std::fs::metadata(&dfx)
-                .with_context(|| {
-                    format!(
-                        "Failed to read file metadata for {}.",
-                        dfx.to_string_lossy()
-                    )
-                })?
-                .permissions();
+            let mut perms = dfx_core::fs::read_permissions(&dfx)?;
             perms.set_mode(EXEC_READ_USER_ONLY_PERMISSION);
-            std::fs::set_permissions(&dfx, perms).with_context(|| {
-                format!("Failed to set file metadata for {}.", dfx.to_string_lossy())
-            })?;
+            dfx_core::fs::set_permissions(&dfx, perms)?;
         }
 
         // atomically install cache version into place
         if force && p.exists() {
-            std::fs::remove_dir_all(&p)
-                .with_context(|| format!("Failed to remove {}.", p.to_string_lossy()))?;
+            dfx_core::fs::remove_dir_all(&p)?;
         }
 
-        if std::fs::rename(&temp_p, &p).is_ok() {
+        if dfx_core::fs::rename(temp_p.as_path(), &p).is_ok() {
             if let Some(b) = b {
                 b.finish_with_message(format!("Version v{} installed successfully.", v));
             }
         } else {
-            std::fs::remove_dir_all(&temp_p).with_context(|| {
-                format!(
-                    "Failed to remove temp binary cache {}.",
-                    temp_p.to_string_lossy()
-                )
-            })?;
+            dfx_core::fs::remove_dir_all(temp_p.as_path())?;
             if let Some(b) = b {
                 b.finish_with_message(format!("Version v{} was already installed.", v));
             }
         }
-
         Ok(p)
     } else {
-        Err(DfxError::new(CacheError::UnknownVersion(v.to_owned())))
+        Err(CacheError::InvalidCacheForDfxVersion(v.to_owned()))
     }
 }
 
-#[context(
-    "Failed to get path to binary '{}' for version '{}'.",
-    binary_name,
-    version
-)]
-pub fn get_binary_path_from_version(version: &str, binary_name: &str) -> DfxResult<PathBuf> {
+pub fn get_binary_path_from_version(
+    version: &str,
+    binary_name: &str,
+) -> Result<PathBuf, CacheError> {
     install_version(version, false)?;
 
     let env_var_name = format!("DFX_{}_PATH", binary_name.replace('-', "_").to_uppercase());
@@ -286,31 +227,22 @@ pub fn get_binary_path_from_version(version: &str, binary_name: &str) -> DfxResu
     Ok(get_bin_cache(version)?.join(binary_name))
 }
 
-#[context("Failed to get binary '{}' for version '{}'.", name, version)]
-pub fn binary_command_from_version(version: &str, name: &str) -> DfxResult<std::process::Command> {
+pub fn binary_command_from_version(
+    version: &str,
+    name: &str,
+) -> Result<std::process::Command, CacheError> {
     let path = get_binary_path_from_version(version, name)?;
     let cmd = std::process::Command::new(path);
 
     Ok(cmd)
 }
 
-#[context("Failed to list cache versions.")]
-pub fn list_versions() -> DfxResult<Vec<Version>> {
+pub fn list_versions() -> Result<Vec<Version>, CacheError> {
     let root = get_bin_cache_root()?;
     let mut result: Vec<Version> = Vec::new();
 
-    for entry in std::fs::read_dir(&root).with_context(|| {
-        format!(
-            "Failed to read bin cache root content at {}.",
-            root.to_string_lossy()
-        )
-    })? {
-        let entry = entry.with_context(|| {
-            format!(
-                "Failed to read an entry in bin cache root at {}.",
-                root.to_string_lossy()
-            )
-        })?;
+    for entry in dfx_core::fs::read_dir(&root)? {
+        let entry = entry.map_err(CacheError::ReadCacheEntryFailed)?;
         if let Some(version) = entry.file_name().to_str() {
             if version.starts_with('_') {
                 // temp directory for version being installed
@@ -318,7 +250,7 @@ pub fn list_versions() -> DfxResult<Vec<Version>> {
             }
             result.push(
                 Version::parse(version)
-                    .with_context(|| format!("Failed to parse version from {}.", version))?,
+                    .map_err(|e| CacheError::MalformedSemverString(version.to_string(), e))?,
             );
         }
     }
@@ -326,17 +258,15 @@ pub fn list_versions() -> DfxResult<Vec<Version>> {
     Ok(result)
 }
 
-pub fn call_cached_dfx(v: &Version) -> DfxResult<ExitStatus> {
+pub fn call_cached_dfx(v: &Version) -> Result<ExitStatus, CacheError> {
     let v = format!("{}", v);
     let command_path = get_binary_path_from_version(&v, "dfx")?;
-    if command_path
-        == std::env::current_exe().context("Failed to get currently running executable.")?
-    {
-        bail!("Invalid cache for version {}.", v)
+    if command_path == dfx_core::foundation::get_current_exe()? {
+        return Err(CacheError::InvalidCacheForDfxVersion(v));
     }
 
-    std::process::Command::new(command_path)
-        .args(std::env::args().skip(1))
-        .status()
-        .map_err(DfxError::from)
+    let mut binding = std::process::Command::new(command_path);
+    let cmd = binding.args(std::env::args().skip(1));
+    let result = dfx_core::process::execute_process(cmd)?;
+    Ok(result)
 }