+++ src/dfx/src/lib/builders/mod.rs
@@ -2,11 +2,11 @@ use crate::config::dfx_version_str;
 use crate::lib::canister_info::CanisterInfo;
 use crate::lib::environment::Environment;
 use crate::lib::error::DfxResult;
-use dfx_core::config::model::dfinity::{Config, Profile};
-
 use crate::lib::models::canister::CanisterPool;
-use crate::lib::provider::get_network_context;
-use crate::util::{self, check_candid_file};
+use crate::util::check_candid_file;
+use dfx_core::config::model::dfinity::{Config, Profile};
+use dfx_core::network::provider::get_network_context;
+use dfx_core::util;
 
 use anyhow::{bail, Context};
 use candid::Principal as CanisterId;
@@ -306,6 +306,25 @@ fn ensure_trailing_newline(s: String) -> String {
         s.push('\n');
         s
     }
+}
+
+/// Set the permission of the given file to be writeable.
+pub fn set_perms_readwrite(file_path: &PathBuf) -> DfxResult<()> {
+    let mut perms = std::fs::metadata(file_path)
+        .with_context(|| {
+            format!(
+                "Failed to read metadata for file {}.",
+                file_path.to_string_lossy()
+            )
+        })?
+        .permissions();
+    perms.set_readonly(false);
+    std::fs::set_permissions(file_path, perms).with_context(|| {
+        format!(
+            "Failed to set permissions for file {}.",
+            file_path.to_string_lossy()
+        )
+    })
 }
 
 type Env<'a> = (Cow<'static, str>, Cow<'a, OsStr>);