+++ src/dfx/src/lib/mod.rs
@@ -1,3 +1,4 @@
+pub mod agent;
 pub mod builders;
 pub mod call_bundled;
 pub mod canister_info;
@@ -24,7 +25,7 @@ pub mod operations;
 pub mod package_arguments;
 pub mod progress_bar;
 pub mod project;
-pub mod provider;
+pub mod replica;
 pub mod replica_config;
 pub mod root_key;
 pub mod sign;