+++ CHANGELOG.md
@@ -2,6 +2,95 @@
 
 # UNRELEASED
 
+## DFX
+
+### fix: make sure assetstorage did file is created as writeable.
+
+### feat: specify id when provisional create canister
+
+When creating a canister on non-mainnet replica, you can now specify the canister ID.
+
+`dfx canister create <CANISTER_NAME> --specified-id <PRINCIPAL>`
+
+`dfx deploy <CANISTER_NAME> --specified-id <PRINCIPAL>`
+
+### fix: update Rust canister template.
+
+`ic-cdk-timers` is included in the dependencies.
+
+### chore: change the default Internet Computer gateway domain to `icp0.io`
+
+By default, DFX now uses the `icp0.io` domain to connect to Internet Computer as opposed to using `ic0.app`. 
+Canisters communicating with `ic0.app` will continue to function nominally.
+
+### feat: --no-asset-upgrade
+
+### fix: `dfx generate` no longer requires non-Motoko canisters to have a canister ID
+Previously, non-Motoko canisters required that the canister was created before `dfx generate` could be called.
+This requirement is now lifted for all canisters except for canisters of type `"motoko"` or canisters that are listed in (transitive) dependencies of a canister of type `"motoko"`.
+It is planned to lift this requirement for Motoko canisters as well, but this requires more work.
+
+### fix: Make `build` field optional in dfx.json
+
+The `build` field in custom canisters was already optional in code, but this fixes it in the schema.
+
+By specifying the `--no-asset-upgrade` flag in `dfx deploy` or `dfx canister install`, you can ensure that the asset canister itself is not upgraded, but instead only the assets themselves are installed.
+
+### feat: Get identity from env var if present
+
+The identity may be specified using the environment variable `DFX_IDENTITY`.
+
+### feat: Add DFX_ASSETS_WASM
+
+Added the ability to configure the WASM module used for assets canisters through the environment variable `DFX_ASSETS_WASM`.
+
+### feat: dfx pull can download wasm
+
+### fix: dfx deploy and icx-asset no longer retry on permission failure
+
+## Asset Canister
+
+Added `validate_take_ownership()` method so that an SNS is able to add a custom call to `take_ownership()`.
+
+Added `is_aliased` field to `get_asset_properties` and `set_asset_properties`.
+
+Added partial support for proposal-based asset updates:
+
+- Batch ids are now stable.  With upcoming changes to support asset updates by proposal,
+  having the asset canister not reuse batch ids will make it easier to verify that a particular
+  batch has been proposed.
+- Added methods:
+  - propose_commit_batch() stores batch arguments for later commit
+  - delete_batch() deletes a batch, intended for use after propose_commit_batch if cancellation needed
+  - compute_evidence() computes a hash ("evidence") over the proposed batch arguments
+
+Added `api_version` endpoint. With upcoming changes we will introduce breaking changes to asset canister's batch upload process. New endpoint will help `ic-asset` with differentiation between API version, and allow it to support all versions of the asset canister. 
+
+## Dependencies
+
+### Frontend canister
+
+- Module hash: 1d170f9732b9c829a1484e7d68d74cf4a5447aa9df8b14df3c373cfda9e29ce0
+- https://github.com/dfinity/sdk/pull/3023
+- https://github.com/dfinity/sdk/pull/3022
+- https://github.com/dfinity/sdk/pull/3021
+- https://github.com/dfinity/sdk/pull/3019
+- https://github.com/dfinity/sdk/pull/3016
+- https://github.com/dfinity/sdk/pull/3015
+- https://github.com/dfinity/sdk/pull/3001
+- https://github.com/dfinity/sdk/pull/2987
+- https://github.com/dfinity/sdk/pull/2982
+
+### Motoko
+
+Updated Motoko to 0.8.4
+
+### ic-ref
+
+Updated ic-ref to 0.0.1-ca6aca90
+
+# 0.13.1
+
 ## Asset Canister
 
 ### Access control